Move checkTerminalActivity to code-server
