+++ patches/sagemaker-ui-post-startup.patch
@@ -2,16 +2,15 @@ Index: sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
 ===================================================================
 --- sagemaker-code-editor.orig/vscode/src/vs/server/node/webClientServer.ts
 +++ sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
-@@ -6,6 +6,8 @@
+@@ -6,6 +6,7 @@
  import { createReadStream, existsSync, writeFileSync } from 'fs';
  import {readFile } from 'fs/promises';
  import { Promises } from 'vs/base/node/pfs';
-+import { spawn } from 'child_process';
-+import * as fs from 'fs';
+++import { spawn } from 'child_process';
+ import * as fs from 'fs';
  import * as path from 'path';
  import * as http from 'http';
- import * as url from 'url';
-@@ -39,6 +41,10 @@ const textMimeType: { [ext: string]: str
+@@ -40,6 +41,10 @@ const textMimeType: { [ext: string]: str
  	'.svg': 'image/svg+xml',
  };
  
@@ -22,7 +21,7 @@ Index: sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
  /**
   * Return an error to the client.
   */
-@@ -102,6 +108,7 @@ export class WebClientServer {
+@@ -142,6 +147,7 @@ export class WebClientServer {
  	private readonly _callbackRoute: string;
  	private readonly _webExtensionRoute: string;
  	private readonly _idleRoute: string;
@@ -30,7 +29,7 @@ Index: sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
  
  	constructor(
  		private readonly _connectionToken: ServerConnectionToken,
-@@ -118,6 +125,7 @@ export class WebClientServer {
+@@ -158,6 +164,7 @@ export class WebClientServer {
  		this._callbackRoute = `${serverRootPath}/callback`;
  		this._webExtensionRoute = `${serverRootPath}/web-extension-resource`;
  		this._idleRoute = '/api/idle';
@@ -38,7 +37,7 @@ Index: sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
  	}
  
  	/**
-@@ -146,6 +154,9 @@ export class WebClientServer {
+@@ -186,6 +193,9 @@ export class WebClientServer {
  				// extension resource support
  				return this._handleWebExtensionResource(req, res, parsedUrl);
  			}
@@ -48,7 +47,7 @@ Index: sagemaker-code-editor/vscode/src/vs/server/node/webClientServer.ts
  
  			return serveError(req, res, 404, 'Not found.');
  		} catch (error) {
-@@ -482,6 +493,41 @@ export class WebClientServer {
+@@ -524,6 +534,41 @@ export class WebClientServer {
  			serveError(req, res, 500, error.message)
  		}
  	}
